+++ src/providers/TransactionsDispatcherContextProvider.tsx
@@ -5,18 +5,20 @@ import { Subscription } from 'rxjs';
 
 import { TransactionsDispatcherContext } from '../contexts';
 import {
-  estimateCrossChainAction,
-  getFirstCrossChainActionByStatus,
-  getCrossChainActionTransactionsByStatus,
   submitEtherspotTransactionsBatch,
   submitWeb3ProviderTransaction,
+  submitEtherspotAndWaitForTransactionHash,
+} from '../utils/transaction';
+import {
+  estimateCrossChainAction,
+  getCrossChainStatusByHash,
+  getFirstCrossChainActionByStatus,
+  getCrossChainActionTransactionsByStatus,
+  updateCrossChainActionTransactionsStatus,
   updateCrossChainActionsTransactionsStatus,
   rejectUnsentCrossChainActionsTransactions,
   filterCrossChainActionsByStatus,
-  updateCrossChainActionTransactionsStatus,
-  getCrossChainStatusByHash,
-  submitEtherspotAndWaitForTransactionHash,
-} from '../utils/transaction';
+} from '../utils/crossChain';
 import { CROSS_CHAIN_ACTION_STATUS } from '../constants/transactionDispatcherConstants';
 import { useEtherspot, useTransactionBuilderModal } from '../hooks';
 import { STORED_GROUPED_CROSS_CHAIN_ACTIONS } from '../constants/storageConstants';
@@ -120,40 +122,44 @@ const TransactionsDispatcherContextProvider = ({ children }: { children: ReactNo
 
     if (estimated?.errorMessage || !estimated) {
       showAlertModal(estimated.errorMessage ?? 'Unable to estimate!');
-      setCrossChainActions((current) => current.map((currentCrossChainAction) => {
+      setCrossChainActions((current) =>
+        current.map((currentCrossChainAction) => {
+          if (crossChainAction.id === currentCrossChainAction.id) {
+            return {
+              ...currentCrossChainAction,
+              destinationCrossChainAction: [
+                updateCrossChainActionTransactionsStatus(
+                  // @ts-ignore
+                  currentCrossChainAction.destinationCrossChainAction[0],
+                  CROSS_CHAIN_ACTION_STATUS.FAILED
+                ),
+              ],
+            };
+          }
+          return currentCrossChainAction;
+        })
+      );
+      return;
+    }
+
+    setCrossChainActions((current) =>
+      current.map((currentCrossChainAction) => {
         if (crossChainAction.id === currentCrossChainAction.id) {
           return {
             ...currentCrossChainAction,
+            estimated,
             destinationCrossChainAction: [
               updateCrossChainActionTransactionsStatus(
                 // @ts-ignore
                 currentCrossChainAction.destinationCrossChainAction[0],
-                CROSS_CHAIN_ACTION_STATUS.FAILED,
+                CROSS_CHAIN_ACTION_STATUS.PENDING
               ),
-            ]
+            ],
           };
         }
         return currentCrossChainAction;
-      }));
-      return;
-    }
-
-    setCrossChainActions((current) => current.map((currentCrossChainAction) => {
-      if (crossChainAction.id === currentCrossChainAction.id) {
-        return {
-          ...currentCrossChainAction,
-          estimated,
-          destinationCrossChainAction: [
-            updateCrossChainActionTransactionsStatus(
-              // @ts-ignore
-              currentCrossChainAction.destinationCrossChainAction[0],
-              CROSS_CHAIN_ACTION_STATUS.PENDING,
-            ),
-          ]
-        };
-      }
-      return currentCrossChainAction;
-    }));
+      })
+    );
 
     const result = await submitEtherspotAndWaitForTransactionHash(
       sdkForPolygonChain as Sdk,
@@ -163,27 +169,29 @@ const TransactionsDispatcherContextProvider = ({ children }: { children: ReactNo
 
     if (result?.errorMessage || !result?.transactionHash?.length) {
       showAlertModal(result.errorMessage ?? 'Unable to send Polygon transaction!');
-      setCrossChainActions((current) => current.map((currentCrossChainAction) => {
-        if (crossChainAction.id === currentCrossChainAction.id) {
-          return {
-            ...currentCrossChainAction,
-            destinationCrossChainAction: [
-              updateCrossChainActionTransactionsStatus(
-                // @ts-ignore
-                currentCrossChainAction.destinationCrossChainAction[0],
-                CROSS_CHAIN_ACTION_STATUS.FAILED,
-              ),
-            ]
-          };
-        }
-        return currentCrossChainAction;
-      }));
+      setCrossChainActions((current) =>
+        current.map((currentCrossChainAction) => {
+          if (crossChainAction.id === currentCrossChainAction.id) {
+            return {
+              ...currentCrossChainAction,
+              destinationCrossChainAction: [
+                updateCrossChainActionTransactionsStatus(
+                  // @ts-ignore
+                  currentCrossChainAction.destinationCrossChainAction[0],
+                  CROSS_CHAIN_ACTION_STATUS.FAILED
+                ),
+              ],
+            };
+          }
+          return currentCrossChainAction;
+        })
+      );
       return;
     }
 
     setCrossChainActions([]);
     showAlertModal('Transaction sent');
-  }
+  };
 
   useEffect(() => {
     let expired = false;
@@ -256,7 +264,7 @@ const TransactionsDispatcherContextProvider = ({ children }: { children: ReactNo
                 transactionsToSend[0],
                 unsentCrossChainAction.chainId,
                 providerAddress,
-                false,
+                false
               )
             : await submitEtherspotTransactionsBatch(
                 isEtherspotPrime(etherspotMode)
@@ -288,7 +296,8 @@ const TransactionsDispatcherContextProvider = ({ children }: { children: ReactNo
 
           const { transactionHash, batchHash } = result;
 
-          setCrossChainActions((current) => current.map((crossChainAction) => {
+          setCrossChainActions((current) =>
+            current.map((crossChainAction) => {
               if (
                 crossChainAction.id !== unsentCrossChainAction.id ||
                 unsentCrossChainAction.multiCallData?.id !== crossChainAction.multiCallData?.id
@@ -314,7 +323,9 @@ const TransactionsDispatcherContextProvider = ({ children }: { children: ReactNo
                   }
 
                   return [...updated, transaction];
-              }, []);
+                },
+                []
+              );
 
               return {
                 ...crossChainAction,
@@ -572,87 +583,98 @@ const TransactionsDispatcherContextProvider = ({ children }: { children: ReactNo
     let updateExpired = false;
 
     const validPendingWeb3CrossChainActions = crossChainActions.filter(
-      (crossChainAction) => crossChainAction.useWeb3Provider &&
+      (crossChainAction) =>
+        crossChainAction.useWeb3Provider &&
         getCrossChainActionTransactionsByStatus(crossChainAction.transactions, CROSS_CHAIN_ACTION_STATUS.PENDING).length
     );
 
     const checkWeb3TransactionStatuses = async () => {
-      await Promise.all(validPendingWeb3CrossChainActions.map(async (crossChainAction) => {
-        const sdkForChain = isEtherspotPrime(etherspotMode)
-          ? await getEtherspotPrimeSdkForChainId(crossChainAction.chainId)
-          : getSdkForChainId(crossChainAction.chainId);
+      await Promise.all(
+        validPendingWeb3CrossChainActions.map(async (crossChainAction) => {
+          const sdkForChain = isEtherspotPrime(etherspotMode)
+            ? await getEtherspotPrimeSdkForChainId(crossChainAction.chainId)
+            : getSdkForChainId(crossChainAction.chainId);
 
-        if (!sdkForChain) return;
+          if (!sdkForChain) return;
 
-        await Promise.all(crossChainAction.transactions.map(async (transaction) => {
-          if (!transaction?.transactionHash || updateExpired) return;
+          await Promise.all(
+            crossChainAction.transactions.map(async (transaction) => {
+              if (!transaction?.transactionHash || updateExpired) return;
 
-          let status: string | undefined;
+              let status: string | undefined;
 
-          if (transaction?.status === CROSS_CHAIN_ACTION_STATUS.PENDING
-            && crossChainAction.type === TRANSACTION_BLOCK_TYPE.KLIMA_STAKE) {
-            try {
-              const crossChainStatus = await getCrossChainStatusByHash(
-                crossChainAction.chainId,
-                crossChainAction.destinationCrossChainAction[0].chainId,
-                transaction.transactionHash,
-              );
+              if (
+                transaction?.status === CROSS_CHAIN_ACTION_STATUS.PENDING &&
+                crossChainAction.type === TRANSACTION_BLOCK_TYPE.KLIMA_STAKE
+              ) {
+                try {
+                  const crossChainStatus = await getCrossChainStatusByHash(
+                    crossChainAction.chainId,
+                    crossChainAction.destinationCrossChainAction[0].chainId,
+                    transaction.transactionHash
+                  );
 
-              if (crossChainStatus?.status
-                && crossChainStatus?.subStatus
-                && crossChainStatus.status.toLowerCase() === 'done'
-                && crossChainStatus.subStatus.toLowerCase() === 'completed') {
-                status = CROSS_CHAIN_ACTION_STATUS.CONFIRMED;
+                  if (
+                    crossChainStatus?.status &&
+                    crossChainStatus?.subStatus &&
+                    crossChainStatus.status.toLowerCase() === 'done' &&
+                    crossChainStatus.subStatus.toLowerCase() === 'completed'
+                  ) {
+                    status = CROSS_CHAIN_ACTION_STATUS.CONFIRMED;
+                  }
+                } catch (e) {
+                  //
+                }
+              } else if (transaction?.status === CROSS_CHAIN_ACTION_STATUS.PENDING) {
+                try {
+                  const submittedTransaction = await (sdkForChain as Sdk).getTransaction({
+                    hash: transaction.transactionHash,
+                  });
+
+                  if (submittedTransaction?.status === TransactionStatuses.Completed) {
+                    status = CROSS_CHAIN_ACTION_STATUS.CONFIRMED;
+                  } else if (submittedTransaction?.status === TransactionStatuses.Reverted) {
+                    status = CROSS_CHAIN_ACTION_STATUS.FAILED;
+                  }
+                } catch (e) {
+                  //
+                }
               }
-            } catch (e) {
-              //
-            }
-          } else if (transaction?.status === CROSS_CHAIN_ACTION_STATUS.PENDING) {
-            try {
-              const submittedTransaction = await (sdkForChain as Sdk).getTransaction({
-                hash: transaction.transactionHash,
-              });
 
-              if (submittedTransaction?.status === TransactionStatuses.Completed) {
-                status = CROSS_CHAIN_ACTION_STATUS.CONFIRMED;
-              } else if (submittedTransaction?.status === TransactionStatuses.Reverted) {
-                status = CROSS_CHAIN_ACTION_STATUS.FAILED;
+              if (!status) return;
+
+              try {
+                setCrossChainActions((current) =>
+                  current.map((currentCrossChainAction) => {
+                    if (crossChainAction.id !== currentCrossChainAction.id) return currentCrossChainAction;
+                    return {
+                      ...currentCrossChainAction,
+                      transactions: currentCrossChainAction.transactions.map((currentTransaction) => {
+                        if (currentTransaction.transactionHash !== transaction.transactionHash) {
+                          return currentTransaction;
+                        }
+                        return { ...currentTransaction, status, finishTimestamp: +new Date() };
+                      }),
+                    };
+                  })
+                );
+
+                setPendingCrossChainActionIds((current) => current.filter((id) => id !== crossChainAction.id));
+              } catch (e) {
+                //
               }
-            } catch (e) {
-              //
-            }
-          }
 
-          if (!status) return;
-
-          try {
-            setCrossChainActions((current) =>
-              current.map((currentCrossChainAction) => {
-                if (crossChainAction.id !== currentCrossChainAction.id) return currentCrossChainAction;
-                return {
-                  ...currentCrossChainAction,
-                  transactions: currentCrossChainAction.transactions.map((currentTransaction) => {
-                    if (currentTransaction.transactionHash !== transaction.transactionHash) {
-                      return currentTransaction;
-                    }
-                    return { ...currentTransaction, status, finishTimestamp: +new Date() };
-                  }),
-                };
-              })
-            );
-
-            setPendingCrossChainActionIds((current) => current.filter((id) => id !== crossChainAction.id));
-          } catch (e) {
-            //
-          }
-
-          // this is dirty fix due wrong initial implementation, do not repeat
-          if (status === CROSS_CHAIN_ACTION_STATUS.CONFIRMED
-            && crossChainAction.type === TRANSACTION_BLOCK_TYPE.KLIMA_STAKE) {
-            await finalizeKlimaCrossChainAction(crossChainAction);
-          }
-        }));
-      }));
+              // this is dirty fix due wrong initial implementation, do not repeat
+              if (
+                status === CROSS_CHAIN_ACTION_STATUS.CONFIRMED &&
+                crossChainAction.type === TRANSACTION_BLOCK_TYPE.KLIMA_STAKE
+              ) {
+                await finalizeKlimaCrossChainAction(crossChainAction);
+              }
+            })
+          );
+        })
+      );
 
       timeout = setTimeout(checkWeb3TransactionStatuses, 3000);
     };
